% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_peak_phase}
\alias{get_peak_phase}
\title{get_peak_phase}
\usage{
get_peak_phase(vert, amp, per, horiz)
}
\arguments{
\item{vert}{Vertical offset}

\item{amp}{Amplitude}

\item{per}{Period}

\item{horiz}{Horizontal offset}
}
\description{
get_peak_phase
}
