% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_RSS}
\alias{get_RSS}
\title{get_RSS}
\usage{
get_RSS(qr_obj, y_mat)
}
\arguments{
\item{qr_obj}{QR decomposition object}

\item{y_mat}{Matrix to use as second argument to qr()}
}
\description{
get_RSS
}
