% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_sine_points}
\alias{get_sine_points}
\title{get_sine_points}
\usage{
get_sine_points(x, vert, amp, per, horiz)
}
\arguments{
\item{x}{Co-ordinates on x-axis}

\item{vert}{Vertical offset}

\item{amp}{Amplitude}

\item{per}{Period}

\item{horiz}{Horizontal offset}
}
\description{
get_sine_points
}
