% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_Fit}
\alias{plot_Fit}
\alias{plot_Fit,sineFitData,ANY-method}
\alias{plot_Fit,sineFitData-method}
\title{plot_Fit}
\usage{
plot_Fit(object, genes, y_lower = NULL, y_upper = NULL, plot_title = NULL)

\S4method{plot_Fit}{sineFitData}(object, genes, y_lower = NULL, y_upper = NULL, plot_title = NULL)
}
\arguments{
\item{object}{A sineFitData object}

\item{genes}{A vector of genes to plot}

\item{y_lower}{Optional lower y-axis limit (defaults to minimum expression or sine wave point)}

\item{y_upper}{Optional upper y-axis limit (defaults to maximum expression or sine wave point)}

\item{plot_title}{Optional custom plot title (defaults to object name and comma-separated list of gene names)}
}
\value{
A ggplot object
}
\description{
plot_Fit
}
