% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_Fit}
\alias{plot_Fit}
\alias{plot_Fit.sf}
\title{plot_Fit}
\usage{
plot_Fit.sf(object, genes, y_lower = NULL, y_upper = NULL, plot_title = NULL)
}
\arguments{
\item{object}{A sineFitData object}

\item{genes}{A vector of genes to plot}

\item{y_lower}{Optional lower y-axis limit}

\item{y_upper}{Optional upper y-axis limit}

\item{plot_title}{Optional plot title}
}
\description{
plot_Fit
}
