% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_PeriodFits}
\alias{plot_PeriodFits}
\alias{plot_PeriodFits,sineFitData,ANY-method}
\alias{plot_PeriodFits,sineFitData-method}
\title{plot_PeriodFits}
\usage{
plot_PeriodFits(
  object,
  gene,
  periods,
  y_lower = NULL,
  y_upper = NULL,
  plot_title = NULL
)

\S4method{plot_PeriodFits}{sineFitData}(
  object,
  gene,
  periods,
  y_lower = NULL,
  y_upper = NULL,
  plot_title = NULL
)
}
\arguments{
\item{object}{A sineFitData object}

\item{gene}{A gene to plot various period fits for}

\item{periods}{A vector of periods to plot. The best fit for each given period is calculated and displayed.
Note that given period(s) need not be in the interval used when generating the inout sineFit object.}

\item{y_lower}{Optional lower y-axis limit (defaults to minimum expression or sine wave point)}

\item{y_upper}{Optional upper y-axis limit (defaults to maximum expression or sine wave point)}

\item{plot_title}{Optional custom plot title (defaults to object name and gene name)}
}
\value{
A ggplot object
}
\description{
plot_PeriodFits
}
