% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_RSquared}
\alias{plot_RSquared}
\alias{plot_RSquared,sineFitData,ANY-method}
\alias{plot_RSquared,sineFitData-method}
\title{plot_RSquared}
\usage{
plot_RSquared(object, gene, plot_adj_rsquared = FALSE)

\S4method{plot_RSquared}{sineFitData}(object, gene, plot_adj_rsquared = FALSE)
}
\arguments{
\item{object}{A sineFitData object}

\item{gene}{Gene name}

\item{plot_adj_rsquared}{Logical to indicate if adjusted R-Squared should be plotted}
}
\description{
plot_RSquared
}
