% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_RSS}
\alias{plot_RSS}
\alias{plot_RSS.sineFit}
\title{plot_RSS}
\usage{
plot_RSS.sineFit(object, gene)
}
\arguments{
\item{object}{A sineFitData object}

\item{gene}{Gene name}
}
\description{
plot_RSS
}
