% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{results}
\alias{results}
\alias{results,sineFitData,ANY-method}
\alias{results,sineFitData-method}
\title{results}
\usage{
results(object)

\S4method{results}{sineFitData}(object)
}
\arguments{
\item{object}{A sinFitData object}
}
\value{
A tibble with fit data and statistics for each gene
}
\description{
results
}
