% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{show_documentsetMethod}
\alias{show_documentsetMethod}
\alias{show,sineFitData,ANY-method}
\alias{show,sineFitData-method}
\title{Show summary of object}
\usage{
\S4method{show}{sineFitData}(object)
}
\arguments{
\item{object}{a sineFitData object}
}
\description{
Show summary of object
}
