% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{sineFit}
\alias{sineFit}
\alias{sineFit,data.frame,data.frame,ANY-method}
\alias{sineFit,data.frame,data.frame-method}
\title{sineFit}
\usage{
sineFit(in_data, rowData, min_per = 24, max_per = 24)

\S4method{sineFit}{data.frame,data.frame}(in_data, rowData, min_per = 24, max_per = 24)
}
\arguments{
\item{in_data}{Data frame with sample IDs and expression values}

\item{rowData}{Data frame with sample IDs, time points and any other covariates}

\item{min_per}{Minimum period to generate fit}

\item{max_per}{Maximum period to generate fit}
}
\value{
sf A sineFitData object
}
\description{
sineFit
}
