% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_Fit}
\alias{plot_Fit}
\alias{plot_Fit,sineFitData,ANY-method}
\alias{plot_Fit,sineFitData-method}
\title{Plot the best fits}
\usage{
plot_Fit(
  object,
  genes,
  y_lower = NULL,
  y_upper = NULL,
  plot_title = NULL,
  gene_cols = NULL
)

\S4method{plot_Fit}{sineFitData}(
  object,
  genes,
  y_lower = NULL,
  y_upper = NULL,
  plot_title = NULL,
  gene_cols = NULL
)
}
\arguments{
\item{object}{A sineFitData object}

\item{genes}{A vector of genes to plot}

\item{y_lower}{Optional lower y-axis limit (defaults to minimum expression or sine wave point)}

\item{y_upper}{Optional upper y-axis limit (defaults to maximum expression or sine wave point)}

\item{plot_title}{Optional custom plot title (defaults to object name and comma-separated list of gene names)}

\item{gene_cols}{Optional vector of (corresponding) colours for genes. Should be at least the same length as \code{genes} or
the default \code{ggplot} palette will be used.}
}
\value{
A ggplot object
}
\description{
Plot the expression and best fitting sine waves for a vector of genes.
}
\details{
This plotting function allows for a limited amount of customisation (title, y-axis limits, colour of genes). Use \code{layer_data()}
from \code{ggplot} on the returned object to obtain the plot data if you wish to create your own cutomised plot.
}
