% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_FStat}
\alias{plot_FStat}
\alias{plot_FStat,sineFitData,ANY-method}
\alias{plot_FStat,sineFitData-method}
\title{plot_FStat}
\usage{
plot_FStat(object, gene, plot_pvalues = FALSE)

\S4method{plot_FStat}{sineFitData}(object, gene, plot_pvalues = FALSE)
}
\arguments{
\item{object}{A sineFitData object}

\item{gene}{Gene name}

\item{plot_pvalues}{Logical to indicate if p-values should be plotted}
}
\value{
A ggplot object
}
\description{
plot_FStat
}
