% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_PeriodFits}
\alias{plot_PeriodFits}
\alias{plot_PeriodFits,sineFitData,ANY-method}
\alias{plot_PeriodFits,sineFitData-method}
\title{Plot the best fits for given periods}
\usage{
plot_PeriodFits(
  object,
  gene,
  periods,
  y_lower = NULL,
  y_upper = NULL,
  plot_title = NULL
)

\S4method{plot_PeriodFits}{sineFitData}(
  object,
  gene,
  periods,
  y_lower = NULL,
  y_upper = NULL,
  plot_title = NULL
)
}
\arguments{
\item{object}{A sineFitData object}

\item{gene}{A gene for which to plot various period fits}

\item{periods}{A vector of periods for which to plot the best fits}

\item{y_lower}{Optional lower y-axis limit (defaults to minimum expression or sine wave point)}

\item{y_upper}{Optional upper y-axis limit (defaults to maximum expression or sine wave point)}

\item{plot_title}{Optional custom plot title (defaults to object name and gene name)}
}
\value{
A ggplot object
}
\description{
Plot the expression and fitted sine waves for a gene and a given vector of period values.
}
\details{
This function is intended to allow visual comparisons of the best fits for different period values. The function calls
\code{sineFit} for the given periods, so it will work even if the given periods were not in the interval used when generating
the input \code{sineFitData} object. Allows for a limited amount of customisation (title and y-axis limits). Use \code{layer_data()}
from \code{ggplot} on the returned object to obtain the plot data if you wish to create your own cutomized plot.
}
