% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{results}
\alias{results}
\alias{results,sineFitData,ANY-method}
\alias{results,sineFitData-method}
\title{Access a table of results for a sineFitData object}
\usage{
results(object)

\S4method{results}{sineFitData}(object)
}
\arguments{
\item{object}{A sineFitData object}
}
\value{
A data frame with fit data and statistics for each gene
}
\description{
Obtain a data frame containing sine wave parameters and linear model statistics for a sineFitData object.
}
\details{
The returned data frame contains data for the best-fitting sine waves and has columns:
\enumerate{
\item \code{Gene:}   the gene name
\item \code{lm_Ftest_Pval:}   the (unadjusted) p-value for the linear F-statistic
\item \code{lm_Ftest_Adj_Pval:}   the p-value for the linear F-statistic adjusted for multiple test correction
\item \code{lm_RSquared:}   the linear R-squared value
\item \code{lm_Adj_RSquared:}   the adjusted linear R-squared value
\item \code{lm_RSS:}   the residual sum of squares for the linear model
\item \code{Base:}   the base (mean) level for the sine
\item \code{Amp:}   the amplitude for the sine
\item \code{Period:}   the period (angular frequency) for the sine
\item \code{Phase_Shift:}   the phase-shift for the sine
\item \code{Peak_Phase:}   the time at which the sine is at it's maximum
}
}
