% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{sineFit}
\alias{sineFit}
\alias{sineFit,data.frame,data.frame,ANY-method}
\alias{sineFit,data.frame,data.frame-method}
\title{sineFit}
\usage{
sineFit(in_data, rowData, min_per = 24, max_per = 24)

\S4method{sineFit}{data.frame,data.frame}(in_data, rowData, min_per = 24, max_per = 24)
}
\arguments{
\item{in_data}{Data frame with sample IDs in first column and gene expression values in subsequent columns}

\item{rowData}{Data frame with (at minimum) sample IDs in 1st column and sample time points in 2nd column.
Additional factors/covariates for the linear model can be added as subsequent columns.}

\item{min_per, }{Minimum period to generate fit}

\item{max_per}{Maximum period to generate fit}
}
\value{
sf A sineFitData object
}
\description{
sineFit
}
