% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{sineFit}
\alias{sineFit}
\alias{sineFit,data.frame,data.frame,ANY-method}
\alias{sineFit,data.frame,data.frame-method}
\title{Fast sinusoidal regression for time-series data using a linearized model}
\usage{
sineFit(in_data, rowData, min_per = 24, max_per = 24)

\S4method{sineFit}{data.frame,data.frame}(in_data, rowData, min_per = 24, max_per = 24)
}
\arguments{
\item{in_data}{Data frame with sample IDs in the 1st column and gene expression values in subsequent columns}

\item{rowData}{Data frame with (at minimum) sample IDs in the 1st column and sample time points in the 2nd column.
Additional factors/covariates for the linear model can be added as subsequent columns.}

\item{min_per}{Minimum period to generate fit}

\item{max_per}{Maximum period to generate fit}
}
\value{
A sineFitData object containing the best fits for each gene.
}
\description{
This function uses a linearized model to calculate the mean expression level, amplitude, period and phase-shift
for the best-fitting sine waves for a given a set of timed gene expression levels.
}
\details{
\code{sineFit} returns the best fit (the fit which minimizes RSS) for each gene within the given time interval \code{[min_per, max_per]}.
RSS is calculated at 5-minute intervals by default. To obtain a fixed period fit, set \code{min_per} and \code{max_per} to the same value.
}
